Added support for existing databases.
Now we can use an already existing database configured in the project settings files as endpoint to write the fixtures.
We need it as we're using a geospatial database and the generation of fixtures fails using an in memory sqlite.
Coudn't make it to work with spatialite so we decided to use our development postgres+postgis instance to generate the fixtures.